+++ aps/tokenizer/__init__.py
@@ -1,3 +1,3 @@
-from .base import WordTokenizer
+from .base import WordTokenizer, ApsTokenizer, TokenizerAbc
 from .subword import SubwordTokenizer
 from .vocab import Vocab