+++ examples/sse/librimix/run.sh
@@ -9,37 +9,95 @@ stage=1
 dataset=librimix
 librimix_data_dir=/mnt/jwu/librimix
 
+# data setup
+sr=16k
+mode=max
+spk=2
+
 # train
-gpu=0
+gpu="0,1,2,3"
 exp=1a
 epochs=100
 batch_size=32
+tensorboard=false
+
+# evaluate
+metric=sisnr
+eval_data=mix_both
 
 . ./utils/parse_options.sh || exit 1
+
+# e.g., 2spk_16k_max
+name=${spk}spk_${sr}_${mode}
+# check sr
+case $sr in
+  "16k" )
+  sr_num=16000
+  ;;
+  "8k" )
+  sr_num=8000
+  ;;
+  * )
+  echo "$0: Unknown sr: $sr" && exit 1
+  ;;
+esac
 
 beg=$(echo $stage | awk -F '-' '{print $1}')
 end=$(echo $stage | awk -F '-' '{print $2}')
 [ -z $end ] && end=$beg
 
 if [ $end -ge 1 ] && [ $beg -le 1 ]; then
-  echo "Stage 1: simulate data ..."
+  echo "Stage 1: simulate LibriMix data ..."
   local/simu_librimix.sh $librimix_data_dir
-  local/prep_data.sh $librimix_data_dir $dataset
+fi
+
+if [ $end -ge 2 ] && [ $beg -le 2 ]; then
+  echo "Stage 2: format LibriMix data ..."
+  local/prep_data.sh --mode $mode --spk $spk --sr $sr \
+    $librimix_data_dir $dataset
 fi
 
 cpt_dir=exp/$dataset/$exp
-if [ $end -ge 2 ] && [ $beg -le 2 ]; then
-  echo "Stage 2: training BSS model ..."
-  ./scripts/train.sh \
+if [ $end -ge 3 ] && [ $beg -le 3 ]; then
+  echo "Stage 3: training BSS model ..."
+  ./scripts/distributed_train.sh \
     --gpu $gpu \
     --seed 666 \
     --epochs $epochs \
     --batch-size $batch_size \
+    --tensorboard $tensorboard \
     --num-workers 16 \
-    --eval-interval -1 \
-    --save-interval -1 \
     --prog-interval 250 \
-    --tensorboard false \
     ss $dataset $exp
   echo "$0: Train model done under $cpt_dir"
+fi
+
+if [ $end -ge 4 ] && [ $beg -le 4 ]; then
+  echo "Stage 4: run BSS ..."
+  for name in dev test; do
+    ./cmd/separate.py \
+      --checkpoint $cpt_dir \
+      --sr $sr_num \
+      --device-id $gpu \
+      $data_dir/$name/${eval_data}.scp \
+      $cpt_dir/bss_${name}_${eval_data}
+  done
+fi
+
+if [ $end -ge 5 ] && [ $beg -le 5 ]; then
+  echo "Stage 5: evaluate ..."
+  spk_list=$(seq $spk)
+  for name in dev test; do
+    eval_dir=bss_${name}_${eval_data}
+    for index in $spk_list; do
+      find $cpt_dir/$eval_dir -name "*.wav" | \
+        awk -v ch=$index -F '/' '{printf("%s sox %s -t wav - remix %d |\n", $NF, $0, ch)}' | \
+        sed "s:.wav::" > $cpt_dir/$eval_dir/s${index}.scp
+    done
+    ./cmd/compute_ss_metric.py \
+      --sr $sr_num \
+      --metric $metric \
+      $data_dir/$name/s1.scp,$data_dir/$name/s2.scp \
+      $cpt_dir/$eval_dir/s2.scp,$cpt_dir/$eval_dir/s2.scp
+  done
 fi